moved ability shortcut objects to their own file
